--- AppApp/Scripts/Entity/Label.swift
@@ -1,9 +1,0 @@
-//
-//  Label.swift
-//  AppApp
-//
-//  Created by Reona Kubo on 2020/03/02.
-//  Copyright © 2020 Reona Kubo. All rights reserved.
-//
-
-import Foundation
